+++ docs/project/changelog.rst
@@ -28,7 +28,7 @@ They may change at any time.
 10.3
 ----
 
-*In development*
+*April 17, 2022*
 
 Backwards-incompatible changes
 ..............................