+++ docs/project/changelog.rst
@@ -25,6 +25,11 @@ fixing regressions shortly after a release.
 Only documented APIs are public. Undocumented APIs are considered private.
 They may change at any time.
 
+10.4
+----
+
+*In development*
+
 10.3
 ----